django_app/requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
elasticsearch==8.15.1
flask==3.0.0
redis==5.0.1
requests==2.31.0
